requirements.txt
dvc[azure]==3.61.0
Flask==3.1.1
numpy==2.2.6
Pillow==11.3.0
gunicorn==23.0.0
opencv-python-headless==4.12.0.88
prometheus-client==0.22.1